+++ assets/appFramework/angular/generate/class/unit/jasmine/{GEN_NAME_SNAKE}.spec.js
@@ -1,7 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL}.
  */
-import { {GEN_NAME_PASCAL} } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}";
+import { {GEN_NAME_PASCAL} } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}";
 
 describe("{GEN_NAME_PASCAL}", () => {
     it("can be created", () => {