+++ assets/appFramework/angular/generate/class/unit/mochachai/{GEN_NAME_SNAKE}.spec.js
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL}.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL} } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}";
+import { {GEN_NAME_PASCAL} } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}";
 
 describe("{GEN_NAME_PASCAL}", () => {
     it("can be created", () => {