+++ assets/appFramework/angular/generate/component/unit/mochachai/{GEN_NAME_SNAKE}.component.spec.js
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL}Component.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Component } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.component";
+import { {GEN_NAME_PASCAL}Component } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.component";
 
 describe("{GEN_NAME_PASCAL}Component", () => {
     it("can be created", () => {