+++ assets/appFramework/angular/generate/directive/unit/jasmine/{GEN_NAME_SNAKE}.directive.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL} directive.
  */
-import { {GEN_NAME_PASCAL}Directive } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.directive";
+import { {GEN_NAME_PASCAL}Directive } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.directive";
 
 describe("{GEN_NAME_PASCAL}Directive", () => {
     it("can be created", () => {