+++ assets/appFramework/angular/generate/directive/unit/mochachai/{GEN_NAME_SNAKE}.directive.spec.ts
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL} directive.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Directive } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.directive";
+import { {GEN_NAME_PASCAL}Directive } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.directive";
 
 describe("{GEN_NAME_PASCAL}Directive", () => {
     it("can be created", () => {