+++ assets/appFramework/angular/generate/enum/unit/jasmine/{GEN_NAME_SNAKE}.enum.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL} enum.
  */
-import { {GEN_NAME_PASCAL}Enum } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.enum";
+import { {GEN_NAME_PASCAL}Enum } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.enum";
 
 describe("{GEN_NAME_PASCAL}Enum", () => {
     it("can be created", () => {