+++ assets/appFramework/angular/generate/enum/unit/mochachai/{GEN_NAME_SNAKE}.enum.spec.ts
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL} enum.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Enum } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.enum";
+import { {GEN_NAME_PASCAL}Enum } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.enum";
 
 describe("{GEN_NAME_PASCAL}Enum", () => {
     it("can be created", () => {