+++ assets/appFramework/angular/generate/guard/unit/mochachai/{GEN_NAME_SNAKE}.guard.spec.js
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL} guard.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Guard } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.guard";
+import { {GEN_NAME_PASCAL}Guard } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.guard";
 
 describe("{GEN_NAME_PASCAL}Guard", () => {
     it("can be created", () => {