+++ assets/appFramework/angular/generate/module/unit/mochachai/{GEN_NAME_SNAKE}.module.spec.js
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL}Module.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Module } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.module";
+import { {GEN_NAME_PASCAL}Module } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.module";
 
 describe("{GEN_NAME_PASCAL}Module", () => {
     it("can be created", () => {