+++ assets/appFramework/angular/generate/pipe/unit/jasmine/{GEN_NAME_SNAKE}.pipe.spec.ts
@@ -1,7 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL} pipe.
  */
-import { {GEN_NAME_PASCAL}Pipe } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.pipe";
+import { {GEN_NAME_PASCAL}Pipe } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.pipe";
 
 describe("{GEN_NAME_PASCAL}Pipe", () => {
     it("can be created", () => {