+++ assets/appFramework/angular/generate/pipe/unit/mochachai/{GEN_NAME_SNAKE}.pipe.spec.js
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL} pipe.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Pipe } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.pipe";
+import { {GEN_NAME_PASCAL}Pipe } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.pipe";
 
 describe("{GEN_NAME_PASCAL}Pipe", () => {
     it("can be created", () => {