+++ assets/appFramework/angular/generate/service/unit/jasmine/{GEN_NAME_SNAKE}.service.spec.js
@@ -1,7 +1,7 @@
 /**
  * Tests for {GEN_NAME_PASCAL} service.
  */
-import { {GEN_NAME_PASCAL}Service } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.service";
+import { {GEN_NAME_PASCAL}Service } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.service";
 
 describe("{GEN_NAME_PASCAL}Service", () => {
     it("can be created", () => {