+++ assets/appFramework/angular/generate/service/unit/mochachai/{GEN_NAME_SNAKE}.service.spec.ts
@@ -2,7 +2,7 @@
  * Tests for {GEN_NAME_PASCAL} service.
  */
 import chai from "chai";
-import { {GEN_NAME_PASCAL}Service } from "{GEN_TEST_ROOT}src/{GEN_SUB_FOLDER}{GEN_NAME_SNAKE}.service";
+import { {GEN_NAME_PASCAL}Service } from "{GEN_TEST_RELATIVE}{GEN_NAME_SNAKE}.service";
 
 describe("{GEN_NAME_PASCAL}Service", () => {
     it("can be created", () => {