+++ assets/gulp/tasks/util/async-util.js
@@ -6,15 +6,10 @@ const util = require("util");
 const fs = require("fs");
 const path = require("path");
 const archiver = require("archiver");
+const streamToPromise = require("stream-to-promise");
 
-function stream (gulpStream, storeMethod) {
-    return new Promise((resolve, reject) => {
-        gulpStream.on("error", reject);
-        gulpStream.on("end", resolve);
-        if (storeMethod) {
-            storeMethod(resolve, reject);
-        }
-    });
+function stream (gulpStream) {
+    return streamToPromise(gulpStream);
 }
 
 async function fileExists (filename) {